+++ xenon-common/src/main/java/com/vmware/xenon/common/http/netty/NettyHttpListener.java
@@ -30,8 +30,12 @@ import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelOption;
 import io.netty.channel.nio.NioEventLoopGroup;
 import io.netty.channel.socket.nio.NioServerSocketChannel;
+import io.netty.handler.codec.http2.Http2SecurityUtil;
+import io.netty.handler.ssl.ApplicationProtocolConfig;
+import io.netty.handler.ssl.ApplicationProtocolNames;
 import io.netty.handler.ssl.SslContext;
 import io.netty.handler.ssl.SslContextBuilder;
+import io.netty.handler.ssl.SupportedCipherSuiteFilter;
 
 import com.vmware.xenon.common.Operation;
 import com.vmware.xenon.common.Operation.SocketContext;
@@ -217,9 +221,21 @@ public class NettyHttpListener implements ServiceRequestListener {
         if (isListening()) {
             throw new IllegalStateException("listener already started");
         }
-        this.sslContext = SslContextBuilder.forServer(
-                new File(certFile), new File(keyFile), keyPassphrase)
-                .build();
+
+        SslContextBuilder builder = SslContextBuilder.forServer(new File(certFile),
+                new File(keyFile), keyPassphrase);
+
+        if (NettyChannelContext.isALPNEnabled()) {
+            builder.ciphers(Http2SecurityUtil.CIPHERS, SupportedCipherSuiteFilter.INSTANCE)
+                    .applicationProtocolConfig(new ApplicationProtocolConfig(
+                            ApplicationProtocolConfig.Protocol.ALPN,
+                            ApplicationProtocolConfig.SelectorFailureBehavior.NO_ADVERTISE,
+                            ApplicationProtocolConfig.SelectedListenerFailureBehavior.ACCEPT,
+                            ApplicationProtocolNames.HTTP_2,
+                            ApplicationProtocolNames.HTTP_1_1));
+        }
+
+        this.sslContext = builder.build();
     }
 
     @Override