+++ xenon-common/src/main/java/com/vmware/xenon/common/http/netty/NettyHttpServerInitializer.java
@@ -13,6 +13,7 @@
 
 package com.vmware.xenon.common.http.netty;
 
+import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.ChannelPipeline;
 import io.netty.channel.socket.SocketChannel;
@@ -33,6 +34,8 @@ import io.netty.handler.codec.http2.HttpToHttp2ConnectionHandlerBuilder;
 import io.netty.handler.codec.http2.InboundHttp2ToHttpAdapter;
 import io.netty.handler.codec.http2.InboundHttp2ToHttpAdapterBuilder;
 import io.netty.handler.logging.LogLevel;
+import io.netty.handler.ssl.ApplicationProtocolNames;
+import io.netty.handler.ssl.ApplicationProtocolNegotiationHandler;
 import io.netty.handler.ssl.SslContext;
 import io.netty.handler.ssl.SslHandler;
 import io.netty.util.AsciiString;
@@ -42,10 +45,40 @@ import com.vmware.xenon.common.ServiceHost.ServiceHostState.SslClientAuthMode;
 import com.vmware.xenon.services.common.ServiceUriPaths;
 
 public class NettyHttpServerInitializer extends ChannelInitializer<SocketChannel> {
+
+    private static class Http2NegotiationHandler extends ApplicationProtocolNegotiationHandler {
+
+        private NettyHttpServerInitializer initializer;
+        private SslHandler sslHandler;
+
+        public Http2NegotiationHandler(NettyHttpServerInitializer initializer,
+                SslHandler sslHandler) {
+            super(ApplicationProtocolNames.HTTP_1_1);
+            this.initializer = initializer;
+            this.sslHandler = sslHandler;
+        }
+
+        @Override
+        protected void configurePipeline(ChannelHandlerContext ctx, String protocol)
+                throws Exception {
+            if (ApplicationProtocolNames.HTTP_2.equals(protocol)) {
+                this.initializer.initializeHttp2Pipeline(ctx.pipeline(), this.sslHandler);
+                return;
+            }
+            if (ApplicationProtocolNames.HTTP_1_1.equals(protocol)) {
+                this.initializer.initializeHttpPipeline(ctx.pipeline(), this.sslHandler);
+                return;
+            }
+            throw new IllegalStateException("Unexpected protocol: " + protocol);
+        }
+    }
+
     public static final String AGGREGATOR_HANDLER = "aggregator";
+    public static final String ALPN_HANDLER = "alpn-handler";
     public static final String HTTP_REQUEST_HANDLER = "http-request-handler";
     public static final String WEBSOCKET_HANDLER = "websocket-request-handler";
     public static final String HTTP1_CODEC = "http1-codec";
+    public static final String HTTP2_HANDLER = "http2-handler";
     public static final String HTTP2_UPGRADE_HANDLER = "http2-upgrade-handler";
     public static final String SSL_HANDLER = "ssl";
 
@@ -91,8 +124,22 @@ public class NettyHttpServerInitializer extends ChannelInitializer<SocketChannel
                 }
             }
             p.addLast(SSL_HANDLER, sslHandler);
+
+            if (NettyChannelContext.isALPNEnabled()) {
+                p.addLast(ALPN_HANDLER, new Http2NegotiationHandler(this, sslHandler));
+                return;
+            }
         }
 
+        initializeHttpPipeline(p, sslHandler);
+    }
+
+    private void initializeHttp2Pipeline(ChannelPipeline pipeline, SslHandler sslHandler) {
+        pipeline.addLast(HTTP2_HANDLER, makeHttp2ConnectionHandler());
+        initializeCommon(pipeline, sslHandler);
+    }
+
+    private void initializeHttpPipeline(ChannelPipeline p, SslHandler sslHandler) {
         // The HttpServerCodec combines the HttpRequestDecoder and the HttpResponseEncoder, and it
         // also provides a method for upgrading the protocol, which we use to support HTTP/2. It
         // also supports a couple other minor features (support for HEAD and CONNECT), which
@@ -103,7 +150,7 @@ public class NettyHttpServerInitializer extends ChannelInitializer<SocketChannel
                 NettyChannelContext.MAX_CHUNK_SIZE, false);
         p.addLast(HTTP1_CODEC, http1_codec);
         if (this.sslContext == null) {
-            // Today we only use HTTP/2 when SSL is disabled
+            // Configure for plaintext upgrade to HTTP/2
             final HttpToHttp2ConnectionHandler connectionHandler = makeHttp2ConnectionHandler();
             UpgradeCodecFactory upgradeCodecFactory = new UpgradeCodecFactory() {
                 @Override
@@ -129,19 +176,21 @@ public class NettyHttpServerInitializer extends ChannelInitializer<SocketChannel
 
         p.addLast(AGGREGATOR_HANDLER,
                 new HttpObjectAggregator(this.responsePayloadSizeLimit));
+        initializeCommon(p, sslHandler);
+    }
+
+    private void initializeCommon(ChannelPipeline p, SslHandler sslHandler) {
         p.addLast(WEBSOCKET_HANDLER, new NettyWebSocketRequestHandler(this.host,
                 ServiceUriPaths.CORE_WEB_SOCKET_ENDPOINT,
                 ServiceUriPaths.WEB_SOCKET_SERVICE_PREFIX));
         p.addLast(HTTP_REQUEST_HANDLER,
                 new NettyHttpClientRequestHandler(this.host, this.listener, sslHandler,
-                this.responsePayloadSizeLimit));
+                        this.responsePayloadSizeLimit));
     }
 
     /**
      * For HTTP/2 we don't have anything as simple as the HttpServerCodec (at least, not in Netty
      * 5.0alpha 2), so we create the equivalent.
-     *
-     * @return
      */
     private HttpToHttp2ConnectionHandler makeHttp2ConnectionHandler() {
         // DefaultHttp2Connection is for client or server. True means "server".